+++ src/api/index.d.ts
@@ -45,10 +45,12 @@ declare module '@cityofzion/neon-js' {
 
 
     //core
-    export function setApiSwitch(newSetting: number): void
-    export function setSwitchFreeze(newSetting: boolean): void
     export function getBalanceFrom(config: apiConfig, api: object): apiConfig
     export function getClaimsFrom(config: apiConfig, api: object): apiConfig
+    export function getRPCEndpointFrom(config: apiConfig, api: object): apiConfig
+    export function getTransactionHistoryFrom(config: apiConfig, api: object): apiConfig
+    export function getWalletDBHeightFrom(config: apiConfig, api: object): apiConfig
+    export function getMaxClaimAmountFrom(config: apiConfig, api: object): apiConfig
     export function createTx(config: apiConfig, txType: string): apiConfig
     export function signTx(config: apiConfig): apiConfig
     export function sendTx(config: apiConfig): apiConfig
@@ -130,6 +132,10 @@ declare module '@cityofzion/neon-js' {
         signingFunction?: (unsigned: Transaction, publicKey: string) => Transaction
       ): Promise<RPCResponse>
     }
+
+    // switch
+    export function setApiSwitch(newSetting: number): void
+    export function setSwitchFreeze(newSetting: boolean): void
   }
   export interface semantic {
     get: {