+++ src/api/neonDB.js
@@ -13,7 +13,7 @@ export const name = 'neonDB'
  * @param {string} net - 'MainNet', 'TestNet', or custom neon-wallet-db URL.
  * @return {string} URL of API endpoint.
  */
-export const getAPIEndpoint = (net) => {
+export const getAPIEndpoint = net => {
   switch (net) {
     case 'MainNet':
       return 'http://api.wallet.cityofzion.io'
@@ -30,18 +30,16 @@ export const getAPIEndpoint = (net) => {
  * @return {Promise<Balance>} Balance of address
  */
 export const getBalance = (net, address) => {
-  log.warn('Balance object expected to change shape in upcoming version')
   const apiEndpoint = getAPIEndpoint(net)
-  return axios.get(apiEndpoint + '/v2/address/balance/' + address)
-    .then((res) => {
-      const bal = new Balance({ net, address: res.data.address })
-      Object.keys(res.data).map((key) => {
-        if (key === 'net' || key === 'address') return
-        bal.addAsset(key, res.data[key])
-      })
-      log.info(`Retrieved Balance for ${address} from neonDB ${net}`)
-      return bal
+  return axios.get(apiEndpoint + '/v2/address/balance/' + address).then(res => {
+    const bal = new Balance({ net, address: res.data.address })
+    Object.keys(res.data).map(key => {
+      if (key === 'net' || key === 'address') return
+      bal.addAsset(key, res.data[key])
     })
+    log.info(`Retrieved Balance for ${address} from neonDB ${net}`)
+    return bal
+  })
 }
 
 /**
@@ -52,7 +50,7 @@ export const getBalance = (net, address) => {
  */
 export const getClaims = (net, address) => {
   const apiEndpoint = getAPIEndpoint(net)
-  return axios.get(apiEndpoint + '/v2/address/claims/' + address).then((res) => {
+  return axios.get(apiEndpoint + '/v2/address/claims/' + address).then(res => {
     const claimData = res.data
     claimData.claims = claimData.claims.map(c => {
       return {
@@ -64,7 +62,26 @@ export const getClaims = (net, address) => {
         value: c.value
       }
     })
+    log.info(`Retrieved Claims for ${address} from neonDB ${net}`)
     return new Claims(claimData)
+  })
+}
+
+/**
+ * Gets the maximum amount of gas claimable after spending all NEO.
+ * @param {string} net - 'MainNet' or 'TestNet'.
+ * @param {string} address - Address to check.
+ * @return {Promise<Fixed8>} An object with available and unavailable GAS amounts.
+ */
+export const getMaxClaimAmount = (net, address) => {
+  const apiEndpoint = getAPIEndpoint(net)
+  return axios.get(apiEndpoint + '/v2/address/claims/' + address).then(res => {
+    log.info(
+      `Retrieved maximum amount of gas claimable after spending all NEO for ${address} from neonDB ${net}`
+    )
+    return new Fixed8(res.data.total_claim + res.data.total_unspent_claim).div(
+      100000000
+    )
   })
 }
 
@@ -73,9 +90,9 @@ export const getClaims = (net, address) => {
  * @param {string} net - 'MainNet' or 'TestNet'.
  * @return {Promise<string>} The URL of the best performing node.
  */
-export const getRPCEndpoint = (net) => {
+export const getRPCEndpoint = net => {
   const apiEndpoint = getAPIEndpoint(net)
-  return axios.get(apiEndpoint + '/v2/network/best_node').then((response) => {
+  return axios.get(apiEndpoint + '/v2/network/best_node').then(response => {
     log.info(`Best node from neonDB ${net}: ${response.data.node}`)
     return response.data.node
   })
@@ -89,10 +106,12 @@ export const getRPCEndpoint = (net) => {
  */
 export const getTransactionHistory = (net, address) => {
   const apiEndpoint = getAPIEndpoint(net)
-  return axios.get(apiEndpoint + '/v2/address/history/' + address).then((response) => {
-    log.info(`Retrieved History for ${address} from neonDB ${net}`)
-    return response.data.history
-  })
+  return axios
+    .get(apiEndpoint + '/v2/address/history/' + address)
+    .then(response => {
+      log.info(`Retrieved History for ${address} from neonDB ${net}`)
+      return response.data.history
+    })
 }
 
 /**
@@ -100,9 +119,9 @@ export const getTransactionHistory = (net, address) => {
  * @param {string} net - 'MainNet' or 'TestNet'.
  * @return {Promise<number>} Current height.
  */
-export const getWalletDBHeight = (net) => {
+export const getWalletDBHeight = net => {
   const apiEndpoint = getAPIEndpoint(net)
-  return axios.get(apiEndpoint + '/v2/block/height').then((response) => {
+  return axios.get(apiEndpoint + '/v2/block/height').then(response => {
     return parseInt(response.data.block_height)
   })
 }
@@ -122,7 +141,7 @@ export const doClaimAllGas = (net, privateKey, signingFunction) => {
   let signedTx // Scope this outside so that all promises have this
   let endpt
   return Promise.all([rpcEndpointPromise, claimsPromise])
-    .then((values) => {
+    .then(values => {
       endpt = values[0]
       const claims = values[1]
       if (claims.length === 0) throw new Error('No claimable gas!')
@@ -133,11 +152,11 @@ export const doClaimAllGas = (net, privateKey, signingFunction) => {
         return unsignedTx.sign(account.privateKey)
       }
     })
-    .then((signedResult) => {
+    .then(signedResult => {
       signedTx = signedResult
       return Query.sendRawTransaction(signedTx).execute(endpt)
     })
-    .then((res) => {
+    .then(res => {
       if (res.result === true) {
         res.txid = signedTx
       } else {
@@ -159,20 +178,24 @@ export const doClaimAllGas = (net, privateKey, signingFunction) => {
 export const doMintTokens = (net, scriptHash, fromWif, neo, gasCost, signingFunction) => {
   log.warn('doMintTokens will be deprecated in favor of doInvoke')
   const account = new Account(fromWif)
-  const intents = [{ assetId: ASSET_ID.NEO, value: neo, scriptHash: scriptHash }]
+  const intents = [
+    { assetId: ASSET_ID.NEO, value: neo, scriptHash: scriptHash }
+  ]
   const invoke = { operation: 'mintTokens', scriptHash, args: [] }
   const rpcEndpointPromise = getRPCEndpoint(net)
   const balancePromise = getBalance(net, account.address)
   let signedTx
   let endpt
   return Promise.all([rpcEndpointPromise, balancePromise])
-    .then((values) => {
+    .then(values => {
       endpt = values[0]
       let balances = values[1]
-      const attributes = [{
-        data: reverseHex(scriptHash),
-        usage: TxAttrUsage.Script
-      }]
+      const attributes = [
+        {
+          data: reverseHex(scriptHash),
+          usage: TxAttrUsage.Script
+        }
+      ]
       const unsignedTx = Transaction.createInvocationTx(balances, intents, invoke, gasCost, { attributes })
       if (signingFunction) {
         return signingFunction(unsignedTx, account.publicKey)
@@ -180,11 +203,11 @@ export const doMintTokens = (net, scriptHash, fromWif, neo, gasCost, signingFunc
         return unsignedTx.sign(account.privateKey)
       }
     })
-    .then((signedResult) => {
+    .then(signedResult => {
       signedTx = signedResult
       return Query.getContractState(scriptHash).execute(endpt)
     })
-    .then((contractState) => {
+    .then(contractState => {
       const attachInvokedContract = {
         invocationScript: '0000',
         verificationScript: contractState.result.script
@@ -192,7 +215,7 @@ export const doMintTokens = (net, scriptHash, fromWif, neo, gasCost, signingFunc
       signedTx.scripts.unshift(attachInvokedContract)
       return Query.sendRawTransaction(signedTx).execute(endpt)
     })
-    .then((res) => {
+    .then(res => {
       if (res.result === true) {
         res.txid = signedTx.hash
       } else {
@@ -216,13 +239,17 @@ export const doSendAsset = (net, toAddress, from, assetAmounts, signingFunction)
   const toAcct = new Account(toAddress)
   const rpcEndpointPromise = getRPCEndpoint(net)
   const balancePromise = getBalance(net, fromAcct.address)
-  const intents = Object.keys(assetAmounts).map((key) => {
-    return { assetId: ASSET_ID[key], value: assetAmounts[key], scriptHash: toAcct.scriptHash }
+  const intents = Object.keys(assetAmounts).map(key => {
+    return {
+      assetId: ASSET_ID[key],
+      value: assetAmounts[key],
+      scriptHash: toAcct.scriptHash
+    }
   })
   let signedTx
   let endpt
   return Promise.all([rpcEndpointPromise, balancePromise])
-    .then((values) => {
+    .then(values => {
       endpt = values[0]
       const balance = values[1]
       const unsignedTx = Transaction.createContractTx(balance, intents)
@@ -232,11 +259,11 @@ export const doSendAsset = (net, toAddress, from, assetAmounts, signingFunction)
         return unsignedTx.sign(fromAcct.privateKey)
       }
     })
-    .then((signedResult) => {
+    .then(signedResult => {
       signedTx = signedResult
       return Query.sendRawTransaction(signedTx).execute(endpt)
     })
-    .then((res) => {
+    .then(res => {
       if (res.result === true) {
         res.txid = signedTx.hash
       } else {